+++ jackson2/src/main/java/io/github/dfa1/typesafe/jackson2/Jackson2Codec.java
@@ -1,7 +1,7 @@
 package io.github.dfa1.typesafe.jackson2;
 
-import io.github.dfa1.typesafe.Answer;
-import io.github.dfa1.typesafe.Question;
+import io.github.dfa1.typesafe.model.Answer;
+import io.github.dfa1.typesafe.model.Question;
 import io.github.dfa1.typesafe.json.JsonCodec;
 import com.fasterxml.jackson.annotation.JsonSubTypes;
 import com.fasterxml.jackson.annotation.JsonTypeInfo;